Added profile as avatar only for onboarding
