+++ pages/api/profile/avatar.ts
@@ -1,17 +1,10 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
 
-import { prisma } from 'db';
-import { getUserS3FilePath, uploadUrlToS3 } from 'lib/aws/uploadToS3Server';
-import { getNFT } from 'lib/blockchain/nfts';
-import * as alchemyApi from 'lib/blockchain/provider/alchemy';
-import log from 'lib/log';
 import { onError, onNoMatch, requireUser } from 'lib/middleware';
-import { sessionUserRelations } from 'lib/session/config';
+import { updateProfileAvatar } from 'lib/profile/updateProfileAvatar';
 import { withSessionRoute } from 'lib/session/withSession';
 import type { UserAvatar } from 'lib/users/interfaces';
-import { InvalidInputError } from 'lib/utilities/errors';
-import { getFilenameWithExtension } from 'lib/utilities/getFilenameWithExtension';
 import type { LoggedInUser } from 'models';
 
 const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
@@ -20,60 +13,9 @@ handler.use(requireUser).put(updateAvatar);
 
 async function updateAvatar(req: NextApiRequest, res: NextApiResponse<LoggedInUser | { error: string }>) {
   const { avatar, avatarTokenId, avatarContract, avatarChain } = req.body as UserAvatar;
-  const { id: userId } = req.session.user;
+  const userId = req.session.user.id;
 
-  let avatarUrl = avatar || null;
-  const updatedTokenId = (avatar && avatarTokenId) || null;
-  const updatedContract = (avatar && avatarContract) || null;
-
-  if (!!updatedContract !== !!updatedTokenId) {
-    throw new InvalidInputError('Invalid avatar data');
-  }
-
-  const isNftAvatar = avatar && updatedTokenId && updatedContract && avatarChain;
-
-  // Provided NFT data
-  if (isNftAvatar) {
-    const wallets = await prisma.userWallet.findMany({
-      where: {
-        userId: req.session.user.id
-      }
-    });
-    const owners = await alchemyApi.getOwners(updatedContract, updatedTokenId, avatarChain);
-    const isOwner = wallets.some((a) => {
-      return owners.find((o) => o.toLowerCase() === a.address.toLowerCase());
-    });
-
-    if (!isOwner) {
-      throw new InvalidInputError('You do not own the selected NFT');
-    }
-
-    const nft = await getNFT(updatedContract, updatedTokenId, avatarChain);
-
-    if (nft.image) {
-      const pathInS3 = getUserS3FilePath({ userId, url: getFilenameWithExtension(nft.image) });
-      try {
-        const { url } = await uploadUrlToS3({ pathInS3, url: nft.image });
-        avatarUrl = url;
-      } catch (e) {
-        log.error('Failed to save avatar', { error: e, pathInS3, url: nft.image, userId });
-        throw new InvalidInputError('Failed to save avatar');
-      }
-    }
-  }
-
-  const user = await prisma.user.update({
-    where: {
-      id: userId
-    },
-    include: sessionUserRelations,
-    data: {
-      avatar: avatarUrl,
-      avatarContract: updatedContract || null,
-      avatarTokenId: updatedTokenId || null,
-      avatarChain: isNftAvatar ? avatarChain : null
-    }
-  });
+  const user = await updateProfileAvatar({ avatar, avatarTokenId, avatarContract, avatarChain, userId });
 
   res.status(200).json(user);
 }
